feat/fix: implement protocol params updates + fix params model
  Somehow, few fields were missing or ignored in the encoder / decoder. Plus, I've also adjusted some of the names to reduce the inconsistency with the ProtocolParamUpdate that lives in Pallas.

+++ crates/amaru-ledger/src/summary/governance/backward_compatibility.rs
@@ -99,13 +99,13 @@ fn last_dormant_period<'a>(
     proposals: BTreeSet<&'a (TransactionPointer, Epoch)>,
 ) -> LastDormantEpoch<'a> {
     let dormant_period = |previous_epoch, current_epoch| {
-        if previous_epoch <= current_epoch + governance_action_lifetime as u64 {
+        if previous_epoch <= current_epoch + governance_action_lifetime {
             None
         } else {
             // We iterate on a BTreeSet in reverse order, so when in dormant periods, the previous
             // epoch is necessarily larger than the current one. And we've just guaranteed there's
             // a gap that is greater than the governance_action_lifetime between both.
-            Some(previous_epoch - current_epoch - governance_action_lifetime as u64)
+            Some(previous_epoch - current_epoch - governance_action_lifetime)
         }
     };
